testCreateNewItemWithNullNameAdd
