+++ src/test/java/school/redrover/GroupJavaExplorersTest.java
@@ -1,12 +1,10 @@
 package school.redrover;
 
 import org.openqa.selenium.By;
-import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebElement;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import school.redrover.runner.BaseTest;
-
 import java.time.Duration;
 import java.util.List;
 
@@ -77,23 +75,6 @@ public class GroupJavaExplorersTest extends BaseTest {
         WebElement errorMessage = getDriver().findElement(By.xpath("//div[@id = 'itemname-invalid']"));
 
         Assert.assertTrue(errorMessage.getText().contains(expectedErrorMessage));
-    }
-
-    @Test
-    public void testCreateNewItemWithNullName() {
-        final String expectedErrorMessage = "» This field cannot be empty, please enter a valid name";
-
-        getDriver().manage().timeouts().implicitlyWait(Duration.ofSeconds(5));
-
-        WebElement buttonCreateItem = getDriver().findElement(By.xpath("//a[@href='/view/all/newJob']"));
-        buttonCreateItem.click();
-
-        WebElement buttonMultibranchPipeline = getDriver().findElement(By.xpath("//span[text() = 'Multibranch Pipeline']"));
-        buttonMultibranchPipeline.click();
-
-        WebElement errorMessage = getDriver().findElement(By.xpath("//div[@id = 'itemname-required']"));
-
-        Assert.assertEquals(errorMessage.getText(), expectedErrorMessage);
     }
 
     @Test
